feat: update accounting and conversion modules to use Decimal for monetary values

+++ src/v4vapp_backend_v2/hive_models/custom_json_data.py
@@ -62,7 +62,7 @@ class KeepsatsTransfer(BaseModel):
 
     from_account: AccNameType = Field("", alias="hive_accname_from")
     to_account: AccNameType = Field("", alias="hive_accname_to")
-    sats: int | None = Field(
+    sats: Decimal | None = Field(
         None,
         ge=0,
         description="The amount of sats being transferred. Not needed if we are sending a fixed amount invoice, used if we are using a lightning address or zero value invoice (used as an upper limit sometimes)",
